Add stargate exporter and improve polygon bridge
