+++ lib/polygon_bridge/constants.py
@@ -1,18 +1,13 @@
 """
-Provide constants to be used in Polygon Bridge exporter
+Provide constants to be used in polygon_bridge exporter
 """
 
 ############## ADDRESSES ##############
 
 POLYGON_BRIDGE = "0xa0c68c638235ee32657e8f720a23cec1bfc77c77".lower()
-POLYGON_ETHER_BRIDGE = "0x8484Ef722627bf18ca5Ae6BcF031c23E6e922B30".lower()
 
 ############## SIGNATURES ##############
 
-# ETH deposit: send ether to polygon bridge
-
-# ETH withdraw: send ether from polygon ether bridge
-
-# ERC20 deposit: send to polygon bridge
-
-# ERC20 withdraw https://etherscan.io/tx/0x94e2ec4450459051758916e15db397853c01db4f9948680e091ea7b55b680242
+LOCKED_ERC20_SIG = "0x4c209b5fc8ad50758f13e2e1088ba56a560dff690a1c6fef26394f4c03821dbe".lower()
+LOCKED_ETHER_SIG = "0x8dbdbe6e06c7c7e1d5b7f27010392ceebff46e3a1ce1a5ede6d10de9dd592e0d".lower()
+EXITED_ETHER_SIG = "0x7fcf532c15f0a6db0bd6d0e038bea71d30d808c7d98cb3bf7268a95bf5081b65".lower()
