+++ lib/stargate/query.py
@@ -1,63 +1,3 @@
-#####This is a new file#####
-"""
-Provide constants to be used in stargate exporter
-"""
-
-############## ADDRESSES ##############
-
-Router = "0x8731d54E9D02c286767d56ac03e8037C07e01e98".lower()
-ETH_Router = "0x150f94B44927F078737562f0fcF3C95c01Cc2376".lower()
-
-############## SIGNATURES ##############
-# swap sig
-SWAP_SIG = "0x34660fc8af304464529f48a778e03d03e4d34bcd5f9b6f0cfbf3cd238c642f7f".lower()
-# send credits sig
-SEND_CREDITS_SIG = "0x6939f93e3f21cf1362eb17155b740277de5687dae9a83a85909fd71da95944e7".lower()
-
-# Swap: user_address -> Router
-# SendCredits: Router -> user_address
-#####This is a new file#####
-"""
-Provide functions for clickhouse query in stargate exporter
-"""
-
-from lib.constants import ETH_EVENTS_TABLE
-
-
-def build_events_query(start_dt, end_dt):
-    """
-    Read events from eth_events_v2 table using stargate contracts
-    """
-
-    query_string = f"""
-    SELECT 
-        tx_hash,
-        contract_addr,
-        args,
-        dt,
-        log_index,
-        CASE 
-            WHEN signature = '{SWAP_SIG}' THEN 'swap'
-            WHEN signature = '{SEND_CREDITS_SIG}' THEN 'send_credits'
-        END as action,
-        CASE 
-            WHEN signature = '{SWAP_SIG}' THEN 'eth'
-            ELSE 'erc20'
-        END as token_type
-    FROM
-        {ETH_EVENTS_TABLE}
-    WHERE
-        dt >= toDateTime('{start_dt}')
-        AND dt < toDateTime('{end_dt}')
-        AND contract_addr IN ['{Router}', '{ETH_Router}']
-        AND signature IN ['{SWAP_SIG}', '{SEND_CREDITS_SIG}']
-    ORDER BY
-        dt DESC,
-        log_index DESC
-    """
-    return query_string
-
-#####This is a new file#####
 """
 Provide process function that could keep the required data in CH table format
 """
@@ -158,33 +98,4 @@ def generate_structured_records(events):
         structured = build_event(event)
         if structured:
             yield structured
-
-#####This is a new file#####
-"""Stargate Exporter Class"""
-import logging
-from lib.exporter import Exporter
-from lib.stargate.query import build_events_query
-from lib.stargate.process import process
-
-class StargateExporter(Exporter):
-    """
-    Stargate Exporter Class, used to export transactions
-    between ethereum and BSC.
-    """
-    def __init__(self):
-        super().__init__()
-        self.name = "stargate"
-
-    def read_records(self):
-        """
-        Read data from clickhouse
-
-        Return: records after clickhouse client execution
-        """
-        event_query = build_events_query(self.start_dt, self.end_dt)
-        event_records = self.read_ch_client.execute(event_query)
-        return event_records
-
-    def run(self):
-        """The main function to run the Stargate exporter"""
